[api] split discretization file: glt part done
