+++ spl/api/__init__.py
@@ -1,9 +1,10 @@
 # -*- coding: UTF-8 -*-
 
+from spl.api import ast
 from spl.api import basic
 from spl.api import discretization
 from spl.api import essential_bc
-from spl.api import ast
+from spl.api import glt
+from spl.api import grid
 from spl.api import printing
 from spl.api import settings
-from spl.api import grid
